landing_page_service/requirements.txt
Django==3.0.3
django-admin-sortable2==0.7.5